[Socket] Move SocketOperation out of Socket_2_1.cs
