+++ mcs/class/System/System.dll.sources
@@ -541,6 +541,7 @@ System.Net.Sockets/SocketException.cs
 System.Net.Sockets/SocketFlags.cs
 System.Net.Sockets/SocketInformation.cs
 System.Net.Sockets/SocketInformationOptions.cs
+System.Net.Sockets/SocketOperation.cs
 System.Net.Sockets/SocketOptionLevel.cs
 System.Net.Sockets/SocketOptionName.cs
 System.Net.Sockets/SocketShutdown.cs
